Hand optimized v0.3 and fixed some bugs in 0.2
Version 0.2 did not correctly work on "[0e]", it exhausted the string
instead of detecting it as a "wrong literal". This is fixed in 0.3

Performance improvements have been made by replacing `switch` constructs
with if .. elsif .. else constructs, as well as in the `trr` trimright
function.

The 'left over characters' error now reports the position in the JSON string.